backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
numpy
scikit-learn
joblib
requests
pandas